+++ src/main/java/org/d2ab/iterator/doubles/RepeatingDoubleIterator.java
@@ -28,9 +28,9 @@ import java.util.NoSuchElementException;
  */
 public class RepeatingDoubleIterator extends UnaryDoubleIterator {
 	private final DoubleIterable iterable;
-	private long times;
+	private int times;
 
-	public RepeatingDoubleIterator(DoubleIterable iterable, long times) {
+	public RepeatingDoubleIterator(DoubleIterable iterable, int times) {
 		super(DoubleIterator.EMPTY);
 		this.iterable = iterable;
 		this.times = times;
